requirements.txt
absl-py==1.4.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiosignal==1.3.2
aiosqlite==0.21.0
altair==5.1.0
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
attrs==23.1.0
babel==2.17.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.3
bleach==6.2.0
blinker==1.6.2
blis==1.2.0
build==1.2.2.post1
cachetools==5.3.1
catalogue==2.0.10
certifi==2023.7.22
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.2.0
chroma-hnswlib==0.7.6
chromadb==0.6.3
click==8.1.7
cloudpathlib==0.21.0
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
comtypes==1.4.11
confection==0.1.5
contourpy==1.1.0
cycler==0.11.0
cymem==2.0.11
databases==0.9.0
dataclasses-json==0.6.7
datasets==3.4.0
debugpy==1.8.13
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.8
distro==1.9.0
durationpy==0.9
ecdsa==0.19.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
evaluate==0.4.3
executing==2.2.0
fastapi==0.115.11
fastjsonschema==2.21.1
filelock==3.17.0
flashtext==2.7
Flask==3.1.1
flatbuffers==23.5.26
fonttools==4.42.1
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.12.0
gast==0.4.0
gensim==4.3.3
gitdb==4.0.10
GitPython==3.1.32
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc0
google-api-python-client==2.168.0
google-auth==2.22.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.69.1
googletrans==4.0.0rc1
greenlet==3.1.1
groq==0.19.0
grpcio==1.57.0
grpcio-status==1.71.0
gTTS==2.5.4
h11==0.9.0
h2==3.2.0
h5py==3.9.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httplib2==0.22.0
httptools==0.6.4
httpx==0.13.3
httpx-sse==0.4.0
huggingface-hub==0.29.2
humanfriendly==10.0
hyperframe==5.2.0
idna==3.4
imageio==2.37.0
importlib-metadata==6.8.0
importlib_resources==6.5.2
ipykernel==6.29.5
ipython==9.0.1
ipython_pygments_lexers==1.1.1
ipywidgets==8.1.7
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.4.14
jaxlib==0.6.1
jedi==0.19.2
Jinja2==3.1.2
jiter==0.8.2
joblib==1.3.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.15
keras==2.15.0
kiwisolver==1.4.5
kubernetes==32.0.1
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.44
langchain-groq==0.2.5
langchain-text-splitters==0.3.6
langcodes==3.5.0
langsmith==0.3.11
language_data==1.3.0
lazy_loader==0.4
libclang==16.0.6
marisa-trie==1.2.1
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.26.1
matplotlib==3.7.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.2
ml-dtypes==0.2.0
mmh3==5.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
murmurhash==1.0.12
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.30.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.8.1
notebook==7.3.2
notebook_shim==0.2.4
numpy==1.23.5
oauthlib==3.2.2
onnx==1.18.0
onnx-tf==1.10.0
onnxruntime==1.21.0
openai==1.65.4
opencv-contrib-python==4.11.0.86
opencv-python==4.8.1.78
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
opt-einsum==3.3.0
optree==0.14.1
orjson==3.10.15
overrides==7.7.0
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
Pillow==10.0.0
platformdirs==4.3.6
posthog==3.19.1
preshed==3.0.9
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.0
proto-plus==1.26.1
protobuf==4.24.2
psutil==7.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
pydeck==0.8.0
Pygments==2.16.1
Pympler==1.0.1
pyparsing==3.0.9
PyPika==0.48.9
pypiwin32==223
pyproject_hooks==1.2.0
pyreadline3==3.5.4
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose==3.5.0
python-json-logger==3.3.0
python-multipart==0.0.20
pyttsx3==2.98
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32==308
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.2.1
referencing==0.30.2
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.5.2
rpds-py==0.10.0
rsa==4.9
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.11.2
seaborn==0.13.2
Send2Trash==1.8.3
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
smart-open==7.1.0
smmap==5.0.0
sniffio==1.3.1
sounddevice==0.5.2
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.38
srsly==2.5.1
stack-data==0.6.3
starlette==0.46.1
streamlit==1.38.0
sympy==1.13.1
tenacity==8.2.3
tensorboard==2.15.2
tensorboard-data-server==0.7.1
tensorflow==2.15.0
tensorflow-addons==0.22.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
terminado==0.18.1
thinc==8.3.4
threadpoolctl==3.5.0
tifffile==2025.3.30
timm==1.0.15
tinycss2==1.4.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.1
torchaudio==2.6.0
torchvision==0.21.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.14.3
transformers==4.33.2
typeguard==2.13.3
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions==4.14.0
tzdata==2023.3
tzlocal==4.3.1
ultralytics==8.3.112
ultralytics-thop==2.0.14
uri-template==1.3.0
uritemplate==4.1.1
urllib3==1.26.16
uuid==1.30
uvicorn==0.34.0
validators==0.21.2
wasabi==1.1.3
watchdog==3.0.0
watchfiles==1.0.4
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==2.3.7
widgetsnbextension==4.0.14
wrapt==1.14.1
xxhash==3.5.0
yarl==1.18.3
zipp==3.16.2
zstandard==0.23.0
